Add a 'fix github' edge function that updates a github username and ensures all repos are provisioned for a user for all of their classes (#329)
In addition to being a better short-term bandaid for any remaining access provisioning issues, this also supports self-service "fix my changed GitHub username." A better fix would detect changed usernames on its own, but this was the fast and probably correct fix for now! Each invocation of the button press that makes any changes is logged, so that eventually we can track down any remaining issues, and if the rate of username changes is non-trivial, might also warrant a revisit.